requirements.txt

### `requirements.txt`

```plaintext
paho-mqtt
pymongo
fastapi
uvicorn
toml
